refactor: prepare ipaddr extension
