+++ Makefile
@@ -31,7 +31,7 @@ compile-linux:
 	gcc -O3 $(LINIX_FLAGS) src/sqlite3-define.c src/define/*.c -o dist/define.so
 	gcc -O3 $(LINIX_FLAGS) src/sqlite3-fileio.c src/fileio/*.c -o dist/fileio.so
 	gcc -O1 $(LINIX_FLAGS) src/sqlite3-fuzzy.c src/fuzzy/*.c -o dist/fuzzy.so
-	gcc -O3 $(LINIX_FLAGS) src/sqlite3-ipaddr.c -o dist/ipaddr.so
+	gcc -O3 $(LINIX_FLAGS) src/sqlite3-ipaddr.c src/ipaddr/*.c -o dist/ipaddr.so
 	gcc -O3 $(LINIX_FLAGS) src/sqlite3-json1.c -o dist/json1.so
 	gcc -O3 $(LINIX_FLAGS) src/sqlite3-math.c -o dist/math.so -lm
 	gcc -O3 $(LINIX_FLAGS) -DPCRE2_CODE_UNIT_WIDTH=8 -DLINK_SIZE=2 -DHAVE_CONFIG_H -DSUPPORT_UNICODE src/sqlite3-regexp.c src/regexp/regexp.c src/regexp/pcre2/*.c -o dist/regexp.so
@@ -66,7 +66,7 @@ compile-macos:
 	gcc -O3 $(MACOS_FLAGS) -I src src/sqlite3-define.c src/define/*.c -o dist/define.dylib
 	gcc -O3 $(MACOS_FLAGS) -I src src/sqlite3-fileio.c src/fileio/*.c -o dist/fileio.dylib
 	gcc -O1 $(MACOS_FLAGS) -I src src/sqlite3-fuzzy.c src/fuzzy/*.c -o dist/fuzzy.dylib
-	gcc -O3 $(MACOS_FLAGS) -I src src/sqlite3-ipaddr.c -o dist/ipaddr.dylib
+	gcc -O3 $(MACOS_FLAGS) -I src src/sqlite3-ipaddr.c src/ipaddr/*.c -o dist/ipaddr.dylib
 	gcc -O3 $(MACOS_FLAGS) -I src src/sqlite3-json1.c -o dist/json1.dylib
 	gcc -O3 $(MACOS_FLAGS) -I src src/sqlite3-math.c -o dist/math.dylib -lm
 	gcc -O3 $(MACOS_FLAGS) -DPCRE2_CODE_UNIT_WIDTH=8 -DLINK_SIZE=2 -DHAVE_CONFIG_H -DSUPPORT_UNICODE -I src src/sqlite3-regexp.c src/regexp/regexp.c src/regexp/pcre2/*.c -o dist/regexp.dylib
@@ -82,7 +82,7 @@ compile-macos-x86:
 	gcc -O3 $(MACOS_FLAGS) -I src src/sqlite3-define.c src/define/*.c -o dist/x86/define.dylib -target x86_64-apple-macos10.12
 	gcc -O3 $(MACOS_FLAGS) -I src src/sqlite3-fileio.c src/fileio/*.c -o dist/x86/fileio.dylib -target x86_64-apple-macos10.12
 	gcc -O1 $(MACOS_FLAGS) -I src src/sqlite3-fuzzy.c src/fuzzy/*.c -o dist/x86/fuzzy.dylib -target x86_64-apple-macos10.12
-	gcc -O3 $(MACOS_FLAGS) -I src src/sqlite3-ipaddr.c -o dist/x86/ipaddr.dylib -target x86_64-apple-macos10.12
+	gcc -O3 $(MACOS_FLAGS) -I src src/sqlite3-ipaddr.c src/ipaddr/*.c -o dist/x86/ipaddr.dylib -target x86_64-apple-macos10.12
 	gcc -O3 $(MACOS_FLAGS) -I src src/sqlite3-json1.c -o dist/x86/json1.dylib -target x86_64-apple-macos10.12
 	gcc -O3 $(MACOS_FLAGS) -I src src/sqlite3-math.c -o dist/x86/math.dylib -target x86_64-apple-macos10.12 -lm
 	gcc -O3 $(MACOS_FLAGS) -DPCRE2_CODE_UNIT_WIDTH=8 -DLINK_SIZE=2 -DHAVE_CONFIG_H -DSUPPORT_UNICODE -I src src/sqlite3-regexp.c src/regexp/regexp.c src/regexp/pcre2/*.c -o dist/x86/regexp.dylib -target x86_64-apple-macos10.12
@@ -98,7 +98,7 @@ compile-macos-arm64:
 	gcc -O3 $(MACOS_FLAGS) -I src src/sqlite3-define.c src/define/*.c -o dist/arm64/define.dylib -target arm64-apple-macos11
 	gcc -O3 $(MACOS_FLAGS) -I src src/sqlite3-fileio.c src/fileio/*.c -o dist/arm64/fileio.dylib -target arm64-apple-macos11
 	gcc -O1 $(MACOS_FLAGS) -I src src/sqlite3-fuzzy.c src/fuzzy/*.c -o dist/arm64/fuzzy.dylib -target arm64-apple-macos11
-	gcc -O3 $(MACOS_FLAGS) -I src src/sqlite3-ipaddr.c -o dist/arm64/ipaddr.dylib -target arm64-apple-macos11
+	gcc -O3 $(MACOS_FLAGS) -I src src/sqlite3-ipaddr.c src/ipaddr/*.c -o dist/arm64/ipaddr.dylib -target arm64-apple-macos11
 	gcc -O3 $(MACOS_FLAGS) -I src src/sqlite3-json1.c -o dist/arm64/json1.dylib -target arm64-apple-macos11
 	gcc -O3 $(MACOS_FLAGS) -I src src/sqlite3-math.c -o dist/arm64/math.dylib -target arm64-apple-macos11 -lm
 	gcc -O3 $(MACOS_FLAGS) -DPCRE2_CODE_UNIT_WIDTH=8 -DLINK_SIZE=2 -DHAVE_CONFIG_H -DSUPPORT_UNICODE -I src src/sqlite3-regexp.c src/regexp/regexp.c src/regexp/pcre2/*.c -o dist/arm64/regexp.dylib -target arm64-apple-macos11
